+++ catalog/app/components/Collaborators/index.ts
@@ -1,0 +1,1 @@
+export * from './Collaborators'
